+++ Assets/Scenes/UITestingScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9b6e6ba64c7c28b4bbfffb5b1a3b873b
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
