+++ Assets/Scripts/InputSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 931dc6b52fe1bf343bad1d83490ba0c4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
