--- Assets/Scripts/UI/GameState.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-
-namespace UI
-{
-    public enum GameState
-    {
-        MainUI, PauseUI
-    }
-}
