--- Assets/Scripts/UI/GameState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8a70c95899925e844a9a546f5dcc48a1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
