--- Assets/Scripts/UI/PauseUIState.cs
@@ -1,10 +1,0 @@
-﻿namespace UI
-{
-    public class PauseUIState: UIState
-    {
-        public override GameState GetGameState()
-        {
-            return GameState.PauseUI;
-        }
-    }
-}
