--- Assets/Scripts/UI/PauseUIState.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6108679f5fcf4692895ba6590fb1da8c
-timeCreated: 1721642432
