--- Assets/Scripts/UI/UIState.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace UI
-{
-    public abstract class UIState : MonoBehaviour
-    {
-        private CanvasGroup _canvasGroup;
-        
-        public abstract GameState GetGameState();
-    } 
-}
-
-
