--- Assets/Scripts/UI/UIState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 99727a82e47184e48a19d5a6710cec90
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
